requirements.txt
awscli==1.34.10
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
docutils==0.16
Flask==3.0.3
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
pyasn1==0.6.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.4
